ocr_api/requirements.txt
grpcio==1.68.1
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
protobuf==5.29.2
